move migration stuff to a separate file
